multiple changes and adds
Update the create_tables.sql file to add some places and users. Add a review section in the modal for the place details, a function to get the place details when authenticated and modify the eventListener in the place.js file.Add un function to display all the places, modify the function showPlaceDetails to an async one, modify the body of the modal to have all the details, handle some errors in the function, add a function displayAllPlaces to display all places in the scripts.js file. Add the link to the index.js file in the index.html file and delete the link to data.js file because bridge to the db, same in the place.html file