Multiple changes and adds
Add a new function to redirect to the home page by clicking on the logo in scripts.js file. Add the new function to all the DOMContent in each js files. Update all the html file for the logo redirection. Add some new styles for the place card on the review page