Make multiple adds and changes
Add all the places images in the folder images. Add some e styles for the place card image. Add the dynamic creation of the place in the review.js file, ditto in the place.js and index.js files. Add an expiration time to the cookie. Add the logout function and update the loginVisibility function to switch between the login button and the logout button. Update the html files with the logout button. Update the database with some amenities