updates multiple files
Update the readme file with the adress of the index page. Add personalize variables for an hypothetical darkmode on all the css files. Delete the data.js file, it was the data before the link with the DB. Create a darkmode function, still on work. Remove the darkmode button in the html files. Change the address in the CORS in the __init__ file of the app